+++ src/Factory.sol
@@ -9,7 +9,6 @@ import "./MintableNFT.sol";
 /// @title Factory
 /// @notice Responsible for creating new auctions, minting the NFTs after a succesful auction, and creating new loans
 contract Factory {
-
     /*//////////////////////////////////////////////////////////////
                                  ADDRESSES
     //////////////////////////////////////////////////////////////*/
@@ -44,7 +43,13 @@ contract Factory {
                                  EVENTS
     //////////////////////////////////////////////////////////////*/
 
-    event AuctionCreated(address indexed owner, uint auctionId, uint csrNftID, uint principalAmount, uint maxRate);
+    event AuctionCreated(
+        address indexed owner,
+        uint256 auctionId,
+        uint256 csrNftID,
+        uint256 principalAmount,
+        uint256 maxRate
+    );
 
     /*//////////////////////////////////////////////////////////////
                                  ERRORS
@@ -53,20 +58,38 @@ contract Factory {
     error InvalidCsrAddress();
     error InvalidMaxRate(uint16 rate);
     error OnlyAuctionCanDeployLoan();
-    
+
     /// @notice Set the required addresses
     /// @param _csrNft Address of the CSR NFT
     /// @param _fixedLoanData Metadata for the fixed loan NFT
     /// @param _borrowerData Metadata for the borrower NFT
-    constructor(address _csrNft, NFTMetadata memory _fixedLoanData, NFTMetadata memory _borrowerData) {
-        if (_csrNft.code.length == 0)
-            revert InvalidCsrAddress();
+    constructor(
+        address _csrNft,
+        NFTMetadata memory _fixedLoanData,
+        NFTMetadata memory _borrowerData
+    ) {
+        if (_csrNft.code.length == 0) revert InvalidCsrAddress();
         csrNft = _csrNft;
-        MintableNFT fixedLoan = new MintableNFT(_fixedLoanData.name, _fixedLoanData.symbol, _fixedLoanData.baseURI, address(this));
-        MintableNFT borrower = new MintableNFT(_fixedLoanData.name, _borrowerData.symbol, _borrowerData.baseURI, address(this));
+        MintableNFT fixedLoan = new MintableNFT(
+            _fixedLoanData.name,
+            _fixedLoanData.symbol,
+            _fixedLoanData.baseURI,
+            address(this)
+        );
+        MintableNFT borrower = new MintableNFT(
+            _fixedLoanData.name,
+            _borrowerData.symbol,
+            _borrowerData.baseURI,
+            address(this)
+        );
         fixedLoanNft = fixedLoan;
         borrowerNft = borrower;
-        loan = new Loan(_csrNft, address(this), address(fixedLoan), address(borrower));
+        loan = new Loan(
+            _csrNft,
+            address(this),
+            address(fixedLoan),
+            address(borrower)
+        );
         auction = new Auction(address(this), _csrNft, address(loan));
     }
 
@@ -74,12 +97,26 @@ contract Factory {
     /// @param _csrNftID Id of the CSR NFT to create the auction for
     /// @param _principalAmount Principal amount of the loan
     /// @param _maxRate Maximum rate that the owner is willing to pay
-    function startAuction(uint _csrNftID, uint _principalAmount, uint16 _maxRate) external returns (uint auctionId) {
-        if (_maxRate > 1000)
-            revert InvalidMaxRate(_maxRate);
-        auctionId = auction.createAuction(msg.sender, _csrNftID, _principalAmount, _maxRate);
+    function startAuction(
+        uint256 _csrNftID,
+        uint256 _principalAmount,
+        uint16 _maxRate
+    ) external returns (uint256 auctionId) {
+        if (_maxRate > 1000) revert InvalidMaxRate(_maxRate);
+        auctionId = auction.createAuction(
+            msg.sender,
+            _csrNftID,
+            _principalAmount,
+            _maxRate
+        );
         ERC721(csrNft).transferFrom(msg.sender, address(auction), _csrNftID);
-        emit AuctionCreated(msg.sender, auctionId, _csrNftID, _principalAmount, _maxRate);
+        emit AuctionCreated(
+            msg.sender,
+            auctionId,
+            _csrNftID,
+            _principalAmount,
+            _maxRate
+        );
     }
 
     /// @notice Function that is called by the auction contract after the auction has ended succesfully to deploy the loan and NFTs
@@ -90,9 +127,15 @@ contract Factory {
     /// @param _borrower Address of the borrower (winner of the auction)
     /// @param _principalAmount The principal amount of the loan
     /// @param _rate The final rate, i.e. the lowest bid during the auction
-    function deployLoan(uint _auctionId, uint _csrNftId, address _lender, address _borrower, uint _principalAmount, uint16 _rate) external {
-        if (msg.sender != address(auction))
-            revert OnlyAuctionCanDeployLoan();
+    function deployLoan(
+        uint256 _auctionId,
+        uint256 _csrNftId,
+        address _lender,
+        address _borrower,
+        uint256 _principalAmount,
+        uint16 _rate
+    ) external {
+        if (msg.sender != address(auction)) revert OnlyAuctionCanDeployLoan();
         fixedLoanNft.mint(_lender, _auctionId);
         borrowerNft.mint(_borrower, _auctionId);
         loan.createLoan(_auctionId, _csrNftId, _principalAmount, _rate);
